Added functionality to use the web page link in a fomod to supply an optional web page.
Also refactored modinfo into multiple files (as the actual types are opaque anyway) and cleaned up the headers a bit

I also changed 'visit on nexus' from the menu not to open the downloads tab because it was a little confusing
