+++ src/CMakeLists.txt
@@ -35,6 +35,11 @@ SET(organizer_SRCS
     modlist.cpp
     modinfodialog.cpp
     modinfo.cpp
+	modinfobackup.cpp
+	modinfoforeign.cpp
+	modinfooverwrite.cpp
+	modinforegular.cpp
+	modinfowithconflictinfo.cpp
     messagedialog.cpp
     mainwindow.cpp
     main.cpp
@@ -125,6 +130,11 @@ SET(organizer_HDRS
     modlist.h
     modinfodialog.h
     modinfo.h
+	modinfobackup.h
+	modinfoforeign.h
+	modinfooverwrite.h
+	modinforegular.h
+	modinfowithconflictinfo.h
     messagedialog.h
     mainwindow.h
     loghighlighter.h
